Split up into two projects to act as static library and program, some changes in visual studio to reflect this.

+++ SubnetCalculator/SubnetCalculator/SubnetMask.cpp
@@ -12,7 +12,7 @@ namespace subcalc {
 	int SubnetMask::count() const {
 		int tmp = 0;
 		for (int i = 0; i < 3; i++)
-			tmp += address[i].count();
+			tmp += (int)address[i].count();
 
 		return tmp;
 	}
